+++ packages/network_info_plus/network_info_plus/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.1
+
+- Add IP v6 to MacOS
+
+## 2.0.0
+
+- Remove deprecated method `registerWith` (of Android v1 embedding)
+
 ## 1.3.0
 
 - IOS: Use `NEHotspotNetwork` on iOS 14+ to get SSID/BSSID.